pyproject.toml
[project]
name = "gladys"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "google-api-python-client",
    "matplotlib",
]
